+++ apps/website/content/docs/changelog.md
@@ -2,6 +2,18 @@
 title: Changelog
 ---
 
+## v1.51.0 (2025-06-03)
+
+### ✨ New
+
+- feat(react-x): add `jsx-no-iife` rule, closes #1112 by @Rel1cx in https://github.com/Rel1cx/eslint-react/pull/1113
+
+### 🐞 Fixes
+
+- fix: fixed `no-direct-set-state-in-use-effect` deferred setState calls detection, closes #1117 by @Rel1cx in https://github.com/Rel1cx/eslint-react/pull/1119
+
+**Full Changelog**: https://github.com/Rel1cx/eslint-react/compare/v1.50.0...v1.51.0
+
 ## v1.50.0 (2025-05-25)
 
 ### 🐞 Fixes